@@ -15,21 +15,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "~13.3.9",
-    "@angular/animations": "~13.3.11",
-    "@angular/cdk": "^13.3.9",
-    "@angular/cli": "~13.3.9",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/compiler-cli": "~13.3.11",
-    "@angular/core": "~13.3.11",
-    "@angular/forms": "~13.3.11",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
+    "@angular-devkit/build-angular": "^15.0.0",
+    "@angular/animations": "^15.0.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/cli": "^15.0.0",
+    "@angular/common": "^15.0.0",
+    "@angular/compiler": "^15.0.0",
+    "@angular/compiler-cli": "^15.0.0",
+    "@angular/core": "^15.0.0",
+    "@angular/forms": "^15.0.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^15.0.0",
+    "@angular/platform-browser-dynamic": "^15.0.0",
+    "@angular/router": "^15.0.0",
     "@types/chart.js": "^2.9.29",
-    "angular-bootstrap-md": "^13.0.0",
     "bootstrap": "^4.6.0",
     "chart.js": "^2.9.4",
     "chartist": "^0.11.4",
@@ -39,7 +38,7 @@
     "quill": "^1.3.6",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
-    "typescript": "~4.6.4",
+    "typescript": "~4.8.4",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
@@ -61,7 +60,7 @@
     "tslint": "~6.1.0"
   },
   "engines": {
-    "node": "16.17.0",
-    "npm": "7.5.3"
+    "node": "18.10.0",
+    "npm": "8.19.2"
   }
 }
