@@ -1,6 +1,6 @@
 /**
  * Стратегии торговли
- * Список всех возможный страгтегий 
+ * Список всех возможный страгтегий
  * (теханализ их источников, собственный теханализ, арбитраж и т.д.)
  */
 export enum StrategyList {
@@ -12,8 +12,6 @@ export enum StrategyList {
 /**
  * Общая модель стратегии
  * @param payload - информация о стратегии
- * todo добавить разные интерфейсы для payload в зависимости от типа стратегии;
- * пока у payload один тип;
  */
 export interface StrategyViewModel {
   value: StrategyList;
